corrections from code review
