+++ prometheus.go
@@ -1,8 +1,8 @@
 package gorillaz
 
 import (
+	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"go.uber.org/zap"
 	"strings"
 	"time"
 
@@ -14,37 +14,32 @@ func (g *Gaz) InitPrometheus(path string) {
 	if !strings.HasPrefix(path, "/") {
 		path = "/" + path
 	}
-	g.prometheusRegistry = prometheus.NewRegistry()
 	Sugar.Infof("Setup Prometheus handler at %s", path)
-	g.Router.Handle(path, promhttp.Handler()).Methods("GET")
+	handler := promhttp.InstrumentMetricHandler(
+		g.prometheusRegistry, promhttp.HandlerFor(g.prometheusRegistry, promhttp.HandlerOpts{}),
+	)
+	g.Router.Handle(path, handler).Methods("GET")
 
 	// export uptime as a prometheus counter
 	upCounter := prometheus.NewCounter(prometheus.CounterOpts{
 		Name: "uptime_sec",
 		Help: "uptime in seconds",
 	})
-	if g.RegisterCollector(upCounter) {
-		go func() {
-			t := time.Tick(time.Second)
-			for {
-				<-t
-				upCounter.Inc()
-			}
-		}()
-	}
+	g.RegisterCollector(upCounter)
+	go func() {
+		t := time.Tick(time.Second)
+		for {
+			<-t
+			upCounter.Inc()
+		}
+	}()
 }
 
 // return true if collector was successfully registered
-func (g *Gaz) RegisterCollector(c prometheus.Collector) bool {
-	if g.prometheusRegistry != nil {
-		err := g.prometheusRegistry.Register(c)
-		if err != nil {
-			Log.Warn("Could not register prometheus collector", zap.Error(err))
-			return false
-		}
-		return true
-	} else {
-		Log.Info("No prometheus registry found")
-		return false
+func (g *Gaz) RegisterCollector(c prometheus.Collector) error {
+	err := g.prometheusRegistry.Register(c)
+	if err != nil {
+		return errors.Wrap(err, "Could not register prometheus collector")
 	}
+	return nil
 }
