--- testConfig/application.properties
@@ -1,1 +1,0 @@
-service.name=gorillazServiceTest
